+++ examples/bloc_counters/lib/app/app_observer.dart
@@ -4,23 +4,29 @@ import 'package:bloc/bloc.dart';
 
 class MyBlocObserver extends BlocObserver {
   @override
-  void onEvent(Bloc bloc, Object? event) {
+  void onEvent(Bloc<dynamic, dynamic> bloc, Object? event) {
     super.onEvent(bloc, event);
 
     log('${bloc.runtimeType}: $event', name: 'onEvent');
   }
 
   @override
-  void onTransition(Bloc bloc, Transition transition) {
+  void onTransition(
+    Bloc<dynamic, dynamic> bloc,
+    Transition<dynamic, dynamic> transition,
+  ) {
     super.onTransition(bloc, transition);
     log('${bloc.runtimeType}: $transition', name: 'onTrasition');
   }
 
   @override
-  void onChange(BlocBase bloc, Change state) {
+  void onChange(
+    BlocBase<dynamic> bloc,
+    Change<dynamic> change,
+  ) {
     if (bloc is Cubit) {
-      super.onChange(bloc, state);
-      log('${bloc.runtimeType}: $state', name: 'onChange');
+      super.onChange(bloc, change);
+      log('${bloc.runtimeType}: $change', name: 'onChange');
     }
   }
 }