+++ examples/bloc_counters/lib/blocs/counter_cubit.dart
@@ -1,19 +1,19 @@
 import 'package:bloc/bloc.dart' show Cubit;
 
 class CounterCubitBase extends Cubit<int> {
-  CounterCubitBase(int value) : super(value);
+  CounterCubitBase(super.value);
 
-  void increment([int value = 1]) async {
-    for (int i = 0; i < value; i++) {
+  Future<void> increment([int value = 1]) async {
+    for (var i = 0; i < value; i++) {
       emit(state + 1);
-      await Future.delayed(const Duration(seconds: 2));
+      await Future<void>.delayed(const Duration(seconds: 2));
     }
   }
 
-  void decrement([int value = 1]) async {
-    for (int i = 0; i < value; i++) {
+  Future<void> decrement([int value = 1]) async {
+    for (var i = 0; i < value; i++) {
       emit(state - 1);
-      await Future.delayed(const Duration(seconds: 2));
+      await Future<void>.delayed(const Duration(seconds: 2));
     }
   }
 }