+++ packages/riverbloc/lib/riverbloc.dart
@@ -13,5 +13,7 @@ import 'package:riverbloc/src/framework.dart';
 
 export 'package:bloc/bloc.dart';
 export 'package:riverpod/riverpod.dart';
+
 export 'src/framework.dart';
+export 'src/listenable_provider_whenable.dart';
 export 'src/unimplemented_provider_error.dart';