+++ packages/riverbloc/test/auto_dispose_test.dart
@@ -6,18 +6,16 @@ import 'helpers/helpers.dart';
 
 void main() {
   group('AutoDispose Provider names', () {
-    test('AutoDisposeBlocProvider.notifier with no name', () {
+    test('AutoDisposeBlocProvider.bloc with no name', () {
       final counterBlocProvider = AutoDisposeBlocProvider<CounterBloc, int>(
         (ref) => CounterBloc(0),
       );
-      expect(counterBlocProvider.notifier.name, isNull);
-      expect(counterBlocProvider.stream.name, isNull);
+      expect(counterBlocProvider.name, isNull);
 
       final counterCubitProvider = BlocProvider.autoDispose<CounterCubit, int>(
         (ref) => CounterCubit(0),
       );
-      expect(counterCubitProvider.notifier.name, isNull);
-      expect(counterCubitProvider.stream.name, isNull);
+      expect(counterCubitProvider.name, isNull);
     });
 
     test('BlocProvider.notifier with name', () {
@@ -25,15 +23,13 @@ void main() {
         (ref) => CounterBloc(0),
         name: 'counterBloc',
       );
-      expect(counterBlocProvider.notifier.name, 'counterBloc.notifier');
-      expect(counterBlocProvider.stream.name, 'counterBloc.stream');
+      expect(counterBlocProvider.name, 'counterBloc');
 
       final counterCubitProvider = BlocProvider.autoDispose<CounterCubit, int>(
         (ref) => CounterCubit(0),
         name: 'counterCubit',
       );
-      expect(counterCubitProvider.notifier.name, 'counterCubit.notifier');
-      expect(counterCubitProvider.stream.name, 'counterCubit.stream');
+      expect(counterCubitProvider.name, 'counterCubit');
     });
   });
 
@@ -44,18 +40,17 @@ void main() {
       final container = ProviderContainer();
       expect(
         () => container.read(provider.bloc),
-        throwsA(isA<ProviderException>()),
+        throwsA(isA<UnimplementedProviderError>()),
       );
 
       try {
         container.read(provider.bloc);
-      } on ProviderException catch (e) {
-        expect(e.exception, isA<UnimplementedProviderError>());
-        final unimplementedProviderError =
-            e.exception as UnimplementedProviderError;
-        expect(unimplementedProviderError.name, 'someName');
       } catch (e) {
-        fail('unexpected exception $e');
+        if (e is UnimplementedProviderError) {
+          expect(e.name, 'someName');
+        } else {
+          fail('unexpected exception $e');
+        }
       }
     });
   });
@@ -100,7 +95,7 @@ void main() {
       expect(bloc1, isNot(same(bloc2)));
     });
 
-    test('listening the provider must keep the bloc', () async {
+    test('listening the provider must refresh the bloc', () async {
       final container = ProviderContainer();
       final counterCubitProvider =
           AutoDisposeBlocProv((ref) => CounterCubit(0));
@@ -117,10 +112,8 @@ void main() {
       verify(() => listener(null, bloc1)).called(1);
 
       final bloc2 = container.refresh(counterCubitProvider.bloc);
-      verify(() => listener(bloc1, bloc2)).called(1);
-
-      final bloc3 = container.refresh(counterCubitProvider.bloc);
-      verify(() => listener(bloc2, bloc3)).called(1);
+      verify(() => listener(null, bloc2)).called(1);
+      expect(bloc1, isNot(equals(bloc2)));
 
       sub.close();
     });
@@ -152,8 +145,8 @@ void main() {
       expect(sub1.read(), 0);
       expect(sub2.read(), 0);
 
-      container.read(counterProvider1.notifier).increment();
-      container.read(counterProvider2.notifier).increment();
+      container.read(counterProvider1.bloc).increment();
+      container.read(counterProvider2.bloc).increment();
       await Future(() {});
       expect(sub1.read(), 1);
       verify(() => listener1(0, 1)).called(1);
@@ -178,107 +171,14 @@ void main() {
       expect(closeCounter1, 0);
       expect(closeCounter2, 2);
 
-      container.read(counterProvider1.notifier).increment();
-      container.read(counterProvider2.notifier).increment();
+      container.read(counterProvider1.bloc).increment();
+      container.read(counterProvider2.bloc).increment();
       await Future(() {});
       expect(closeCounter1, 0);
       expect(closeCounter2, 3);
 
       expect(container.read(counterProvider1), 2);
       expect(container.read(counterProvider2), 0);
-
-      await Future(() {});
-      expect(closeCounter1, 0);
-      expect(closeCounter2, 4);
-    });
-
-    test('BlocProvider.stream', () async {
-      var closeCounter1 = 0;
-      var closeCounter2 = 0;
-      void onClose1() => closeCounter1++;
-      void onClose2() => closeCounter2++;
-
-      final listener1 = Listener<AsyncValue<int>>();
-      final listener2 = Listener<AsyncValue<int>>();
-
-      final counterProvider1 = BlocProvider<CounterCubit, int>(
-        (ref) => CounterCubit(0, onClose: onClose1),
-      );
-      final counterProvider2 = BlocProvider.autoDispose<CounterCubit, int>(
-        (ref) => CounterCubit(0, onClose: onClose2),
-      );
-
-      final container = ProviderContainer();
-
-      final sub1 = container.listen<AsyncValue<int>>(
-        counterProvider1.stream,
-        listener1,
-      );
-
-      final sub2 = container.listen<AsyncValue<int>>(
-        counterProvider2.stream,
-        listener2,
-      );
-
-      expect(sub1.read(), equals(const AsyncLoading<int>()));
-      expect(sub2.read(), equals(const AsyncLoading<int>()));
-
-      container.read(counterProvider1.notifier).increment();
-      container.read(counterProvider2.notifier).increment();
-      await Future(() {});
-      expect(sub1.read(), const AsyncData(1));
-      verify(() => listener1(const AsyncLoading(), const AsyncData(1)))
-          .called(1);
-      expect(sub2.read(), const AsyncData(1));
-      verify(() => listener2(const AsyncLoading(), const AsyncData(1)))
-          .called(1);
-
-      container.read(counterProvider1.notifier).increment();
-      container.read(counterProvider2.notifier).increment();
-      await Future(() {});
-      expect(sub1.read(), const AsyncData(2));
-      verify(() => listener1(const AsyncData(1), const AsyncData(2))).called(1);
-      expect(sub2.read(), const AsyncData(2));
-      verify(() => listener2(const AsyncData(1), const AsyncData(2))).called(1);
-
-      verifyNoMoreInteractions(listener1);
-      verifyNoMoreInteractions(listener2);
-
-      expect(closeCounter1, 0);
-      expect(closeCounter2, 0);
-      sub1.close();
-      sub2.close();
-      await Future(() {});
-      expect(closeCounter1, 0);
-      expect(closeCounter2, 1);
-
-      expect(
-        container.read(counterProvider1.stream),
-        const AsyncData(2),
-      );
-      expect(
-        container.read(counterProvider2.stream),
-        const AsyncLoading<int>(),
-      );
-
-      await Future(() {});
-      expect(closeCounter1, 0);
-      expect(closeCounter2, 2);
-
-      container.read(counterProvider1.notifier).increment();
-      container.read(counterProvider2.notifier).increment();
-      await Future(() {});
-      expect(closeCounter1, 0);
-      expect(closeCounter2, 3);
-
-      expect(
-        container.read(counterProvider1.stream),
-        const AsyncData(3),
-      );
-      expect(
-        container.read(counterProvider2.stream),
-        const AsyncLoading<int>(),
-      );
 
       await Future(() {});
       expect(closeCounter1, 0);
@@ -295,13 +195,11 @@ void main() {
       final counterProvider1 = AutoDisposeBlocProvider<CounterCubit, int>(
         (ref) => CounterCubit(0, onClose: onClose1),
       );
-      final counterProvider2 = BlocProvider.autoDispose<CounterCubit, int>(
-        (ref) => CounterCubit(5, onClose: onClose2),
-      );
 
       final container = ProviderContainer(
         overrides: [
-          counterProvider1.overrideWithProvider(counterProvider2),
+          counterProvider1
+              .overrideWith((ref) => CounterCubit(5, onClose: onClose2)),
         ],
       );
 
@@ -309,7 +207,7 @@ void main() {
       final sub = container.listen<int>(counterProvider1, listener);
 
       expect(sub.read(), 5);
-      container.read(counterProvider1.notifier).increment();
+      container.read(counterProvider1.bloc).increment();
       await Future(() {});
       expect(sub.read(), 6);
 
@@ -331,15 +229,7 @@ void main() {
       final container = ProviderContainer();
       final counterCubit = container.read(counterCubitProvider.bloc);
 
-      expect(counterCubit, isA<BlocBase>());
-    });
-
-    test('BlocProvider.bloc equals BlocProvider.notifier', () {
-      final container = ProviderContainer();
-      final bloc = container.read(counterCubitProvider.bloc);
-      final notifier = container.read(counterCubitProvider.notifier);
-
-      expect(bloc, equals(notifier));
+      expect(counterCubit, isA<BlocBase<int>>());
     });
   });