+++ packages/riverbloc/test/riverbloc_test.dart
@@ -3,16 +3,15 @@ import 'package:test/test.dart';
 
 import 'helpers/helpers.dart';
 
-final counterProvider = BlocProv((ref) => CounterBloc(0));
+final counterProvider = BlocProv<CounterBloc>((ref) => CounterBloc(0));
 
-final counterCubitProvider = BlocProv((ref) => CounterCubit(0));
+final counterCubitProvider = BlocProv<CounterCubit>((ref) => CounterCubit(0));
 
 void main() {
   group('Provider names', () {
     test('BlocProvider.bloc with no name', () {
       final counterBlocProvider = BlocProv((ref) => CounterBloc(0));
-      expect(counterBlocProvider.bloc.name, isNull);
-      expect(counterBlocProvider.stream.name, isNull);
+      expect(counterBlocProvider.name, isNull);
     });
 
     test('BlocProvider.bloc with name', () {
@@ -20,8 +19,7 @@ void main() {
         (ref) => CounterBloc(0),
         name: 'counter',
       );
-      expect(counterBlocProvider.bloc.name, 'counter.notifier');
-      expect(counterBlocProvider.stream.name, 'counter.stream');
+      expect(counterBlocProvider.name, 'counter');
     });
   });
 
@@ -31,36 +29,27 @@ void main() {
       final container = ProviderContainer();
       expect(
         () => container.read(provider.bloc),
-        throwsA(isA<ProviderException>()),
+        throwsA(isA<UnimplementedProviderError>()),
       );
 
       try {
         container.read(provider.bloc);
-      } on ProviderException catch (e) {
-        expect(e.exception, isA<UnimplementedProviderError>());
-        final unimplementedProviderError =
-            e.exception as UnimplementedProviderError;
-        expect(unimplementedProviderError.name, 'someName');
       } catch (e) {
-        fail('unexpected exception $e');
+        if (e is UnimplementedProviderError) {
+          expect(e.name, 'someName');
+        } else {
+          fail('unexpected exception $e');
+        }
       }
     });
   });
 
-  group('BlocProvider.notifier', () {
-    test('BlocProvider.notifier gets BlocBase Object', () {
+  group('BlocProvider.bloc', () {
+    test('BlocProvider.bloc gets BlocBase Object', () {
       final container = ProviderContainer();
-      final counterCubit = container.read(counterCubitProvider.notifier);
+      final counterCubit = container.read(counterCubitProvider.bloc);
 
       expect(counterCubit, isA<CounterCubit>());
-    });
-
-    test('BlocProvider.notifier equals BlocProvider.bloc', () {
-      final container = ProviderContainer();
-      final bloc = container.read(counterCubitProvider.bloc);
-      final notifier = container.read(counterCubitProvider.notifier);
-
-      expect(bloc, equals(notifier));
     });
   });
 
@@ -184,56 +173,26 @@ void main() {
 
     test('BlocProvider override with provider', () async {
       final counterBloc = CounterBloc(3);
-      final counterProvider2 =
-          BlocProvider<CounterBloc, int>((ref) => counterBloc);
       final container = ProviderContainer(
         overrides: [
-          counterProvider.overrideWithProvider(counterProvider2),
+          counterProvider.overrideWith((ref) => counterBloc),
         ],
       );
 
       expect(container.read(counterProvider.bloc), equals(counterBloc));
-      expect(container.read(counterProvider2.bloc), equals(counterBloc));
 
       expect(counterBloc.state, 3);
       expect(container.read(counterProvider), 3);
-      expect(container.read(counterProvider2), 3);
 
-      container.read(counterProvider2.bloc).add(Incremented());
+      container.read(counterProvider.bloc).add(Incremented());
       await Future(() {});
       expect(counterBloc.state, 4);
       expect(container.read(counterProvider), 4);
-      expect(container.read(counterProvider2), 4);
 
       container.read(counterProvider.bloc).add(Incremented());
       await Future(() {});
       expect(counterBloc.state, 5);
       expect(container.read(counterProvider), 5);
-      expect(container.read(counterProvider2), 5);
-    });
-
-    test('BlocStateProvider override with value', () async {
-      final bloc2 = CounterBloc(3);
-      final container = ProviderContainer(
-        overrides: [
-          counterProvider.overrideWithValue(bloc2),
-        ],
-      );
-
-      expect(container.read(counterProvider.bloc), equals(bloc2));
-
-      expect(container.read(counterProvider), 3);
-      expect(bloc2.state, 3);
-
-      bloc2.add(Incremented());
-      await Future(() {});
-      expect(container.read(counterProvider), 4);
-      expect(bloc2.state, 4);
-
-      container.read(counterProvider.bloc).add(Incremented());
-      await Future(() {});
-      expect(container.read(counterProvider), 5);
-      expect(bloc2.state, 5);
     });
   });
 
@@ -317,12 +276,9 @@ void main() {
       final pod = BlocProvider<CounterCubit, int>((ref) => CounterCubit(5));
       final container = ProviderContainer();
 
-      expect(container.read(pod.stream), equals(const AsyncLoading<int>()));
-
       container.read(pod.bloc).increment();
       await Future(() {});
 
-      expect(container.read(pod.stream), equals(const AsyncData(6)));
       expect(container.read(pod), 6);
     });
 
@@ -332,58 +288,23 @@ void main() {
       );
       final container = ProviderContainer();
 
-      expect(container.read(pod.stream), equals(const AsyncLoading<int?>()));
       expect(container.read(pod), isNull);
 
       container.read(pod.bloc).increment();
       await Future(() {});
 
-      expect(container.read(pod.stream), equals(const AsyncData<int?>(0)));
       expect(container.read(pod), 0);
-    });
-
-    test('BlocProvider overridden with provider', () {
-      final counterCubit = CounterCubit(3);
-      final counterProvider2 =
-          BlocProvider<CounterCubit, int>((ref) => counterCubit);
-      final container = ProviderContainer(
-        overrides: [
-          counterCubitProvider.overrideWithProvider(counterProvider2),
-        ],
-      );
-
-      expect(container.read(counterCubitProvider.bloc), counterCubit);
-      expect(container.read(counterCubitProvider), 3);
     });
 
     test('BlocProvider overridden with value', () {
       final counterCubit = CounterCubit(5);
       final container = ProviderContainer(
         overrides: [
-          counterCubitProvider.overrideWithValue(counterCubit),
+          counterCubitProvider.overrideWith((ref) => counterCubit),
         ],
       );
       expect(container.read(counterCubitProvider), 5);
       expect(container.read(counterCubitProvider.bloc).state, 5);
-    });
-  });
-
-  group('BlocProvider.setupOverride', () {
-    test('override', () {
-      final cubit2 = CounterCubit(3);
-
-      final counterCubitProvider2 = BlocProvider<CounterCubit, int>(
-        (ref) => cubit2,
-        name: 'cubit2',
-      );
-
-      final override =
-          counterCubitProvider.overrideWithProvider(counterCubitProvider2);
-      expect(override, isA<ProviderOverride>());
-
-      final container = ProviderContainer(overrides: [override]);
-      final cubit = container.read(counterCubitProvider.bloc);
-      expect(cubit, equals(cubit2));
     });
   });